refactor(server): move some tasks to services
