+++ internal/spotifysync/playlist.go
@@ -179,49 +179,3 @@ func (c *client) playlistCoverSync(ctx context.Context, user model.User) error {
 		}
 	}))
 }
-
-func (c *client) playlistRemoveDuplicates(ctx context.Context, user model.User) error {
-	// Spotify's API will remove every instance of a track in a playlist
-	// That's why we first need to make an API call to delete them
-	// Which deleted all instances
-	// And then add them back so that we dont lose them.
-	playlists, err := c.playlist.GetDuplicateTracksByUser(ctx, user.ID)
-	if err != nil {
-		return err
-	}
-
-	for i := range playlists {
-		if playlists[i].SnapshotID == "" {
-			continue
-		}
-
-		unique := utils.SliceUniqueFunc(playlists[i].Duplicates, func(t model.Track) string { return t.SpotifyID })
-		filtered := utils.SliceFilter(unique, func(t model.Track) bool { return t.SpotifyID != "" })
-
-		// Remove them all
-		if err := spotifyapi.C.PlaylistDeleteTrackAll(ctx, user, playlists[i].SpotifyID, playlists[i].SnapshotID, filtered); err != nil {
-			return err
-		}
-
-		// Add them all back
-		if err := spotifyapi.C.PlaylistPostTrackAll(ctx, user, playlists[i].SpotifyID, filtered); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (c *client) PlaylistCreate(ctx context.Context, user model.User, playlist *model.Playlist) error {
-	// TODO: Implement
-	return nil
-}
-
-func (c *client) PlaylistAddTracks(ctx context.Context, user model.User, playlist model.Playlist, tracks []model.Track) error {
-	return spotifyapi.C.PlaylistPostTrackAll(ctx, user, playlist.SpotifyID, tracks)
-}
-
-func (c *client) PlaylistDelete(ctx context.Context, user model.User, playlist model.Playlist) error {
-	// TODO: Implement
-	return nil
-}
